+++ platform/esp32/src/startFunction.c
